requirements.txt
coverage==4.5.3
gitpython==2.1.11
Naked==0.1.31
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
pytest==4.4.1
